fix: safety zone stops killing autonomous navigation
Relay LIDAR safety was sending stop() on teleop during move_base
navigation, overriding the robot's own obstacle avoidance (teleop
has higher mux priority than nav). Now checks navStatus==601 and
skips stop during autonomous navigation.

Also fixes Flutter compile errors:
- int→double for battery in platform/simple transport
- Random() constructor in adaptive_transport_test

+++ flutter_app/lib/core/simple_transport.dart
@@ -125,12 +125,12 @@ class SimpleTransport extends ChangeNotifier {
     // Listen for status updates
     _grpcStatusSub = _grpc!.robotStatus.listen((status) {
       _navStatus = status.navStatus;
-      _battery = status.battery;
+      _battery = status.battery.toDouble();
       _currentGoal = status.navGoal;
 
       _statusController.add(RobotStatusUpdate(
         navStatus: status.navStatus,
-        battery: status.battery,
+        battery: status.battery.toDouble(),
         currentGoal: status.navGoal,
         x: status.pose.x,
         y: status.pose.y,
